rename theme context directory
